pyproject.toml
[project]
name = "MashUp"
version = "0.1.0"
description = "A simple Dash dashboard application"
requires-python = ">=3.8"
dependencies = [
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

